Most of the recipes forms are added.

+++ src/app/shopping-list/shopping-list-editor/shopping-list-editor.component.ts
@@ -38,6 +38,7 @@ export class ShoppingListEditorComponent implements OnInit, OnDestroy {
   }
 
   deleteItem() {
+    if (!this.editMode) return;
     this.ingredientService.deleteIngredient(this.ingredient.id);
     this.clearForm();
   }
